requirements.txt
streamlit
speechrecognition
pydub
google-generativeai
elevenlabs
moviepy~=1.0.3
python-dotenv

